[gradle] Refactor resources configuration
